Get echo_server working with MaaXBoard - exhibiting kernel halt
Strange behaviour - we are using the wrong UART (0x3089000 instead
of 0x3086000 for the MaaxBoard) but that shouldn't matter, we should
just get a virtual memory fault from the kernel. But instead, in debug
mode, we see the the kernel halting. The log has been committed as well
for future debugging.