checkpoint: More progress/fixes/debugging
We can now send out packets and receive responses and pass them
up to the 'client' which in this case is the ARP component I believe.

Still don't quite understanding everything that is going - I am still
likely misunderstanding something in the driver.

Looking at the QEMU logs for network traffic, we can see in wireshark
that a DHCP discover occurs, then a DHCP offer, then nothing. Then
the system seems to continously do discovers and offers.

It seems like the ARP is not getting the kind of packets it expects,
I am not sure why the significance of this since I don't know much
about networking.

Now the best thing to do is probably attach wireshark to a working
driver on hardware that we have and see what we might be missing.